backend/ml_engine/requirements.txt
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.2.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scipy>=1.10.0
